🚸 Improve BookPage to server component
