+++ src/app/book/[id]/page.tsx
@@ -1,7 +1,31 @@
-import BookPage from '@/components/book/BookPage'
+import { getBookInfo } from '@/apis/book'
+import BookInfo from '@/components/book/BookInfo'
+import BookPageButtons from '@/components/book/BookPageButtons'
+import ReadingInfo from '@/components/book/ReadingInfo'
+import PageTitle from '@/components/common/PageTitle'
+import { Suspense } from 'react'
+import Loading from './Loading'
 
-function page({ params }: { params: { id: string } }) {
-  return <BookPage id={params.id} />
+async function page({ params }: { params: { id: string } }) {
+  const { id } = params
+  const info = await getBookInfo({ isbn: id })
+
+  return (
+    <Suspense fallback={<Loading />}>
+      <PageTitle route={info.title} />
+      <ReadingInfo id={id} title={info.title} cover={info.cover} />
+      <BookInfo
+        author={info.author}
+        publisher={info.publisher}
+        pubdate={info.pubdate}
+        desc={info.desc}
+        category={info.category}
+        isbn={id}
+        price={info.price}
+      />
+      <BookPageButtons id={id} />
+    </Suspense>
+  )
 }
 
 export default page
